requirements.txt
requests
langchain
langchain-groq
langchain-core
python-dotenv
